# routes/api.yaml
openapi: 3.0.0
servers:
  - url: /api
info:
  title: basejs API
  version: 0.0.1
  description: Documentation of the RESTful API exposed by basejs
security:
  - JWT: []
paths:
  /auth/token:
    post:
      tags:
        - Authentication
      summary: Create a JWT based on user credentials
      security: []
      operationId: loginWithCredentials
      x-eov-operation-handler: auth
      responses:
        '200':
          description: New generated JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT'
        default:
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
        description: User credentials
        required: true
    delete:
      tags:
        - Authentication
      summary: Log out a user to invalidate existing refresh tokens
      operationId: logout
      x-eov-operation-handler: auth
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/refresh/token:
    post:
      tags:
        - Authentication
      summary: Create a JWT based on a refresh token
      security: []
      operationId: loginWithRefreshToken
      x-eov-operation-handler: auth
      responses:
        '200':
          description: New generated JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT'
        default:
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
        description: refresh token
        required: true
  /auth/social/token:
    post:
      tags:
        - Authentication
      summary: Create a JWT based on social provider token
      security: []
      operationId: loginWithSocialToken
      x-eov-operation-handler: auth
      responses:
        '200':
          description: New generated JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT'
        default:
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - token
              properties:
                provider:
                  type: string
                token:
                  type: string
        description: Social provider and its access token
        required: true
...
components:
  responses:
    NoContentResponse:
      description: OK
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: Bearer
  schemas:
...
    Error:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
    JWT:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string